Split tboplayer.py into different component files Refactor translator code
